requirements.txt
flask
mindsdb_sdk
waitress
flask_jwt_extended
flask-swagger-ui
python-dotenv
psycopg2-binary==2.9.7
flask_cors